+++ src/core/server/api/consts/constDefaultSystems.ts
@@ -1,5 +1,7 @@
 import { DefaultDeathSystem } from '@AthenaServer/systems/defaults/death';
+import { DefaultWeatherSystem } from '@AthenaServer/systems/defaults/weather';
 
 export const defaultSystemsConst = {
     death: DefaultDeathSystem,
+    weather: DefaultWeatherSystem,
 };